[~] refactor code and move some functionla for Article to user
