+++ app/controllers/users/articles_controller.rb
@@ -1,0 +1,52 @@
+module Users
+  class ArticlesController < ApplicationController
+  before_action :set_article, only: [:show, :edit, :update, :destroy]
+  before_action :authenticate_user!
+
+  def index
+    @articles = Article.all
+  end
+
+  def show
+  end
+
+  def new
+    @article = Article.new
+  end
+
+  def edit
+  end
+
+  def create
+    @article = Article.new(article_params)
+
+    if @article.save
+      redirect_to [:users, @article], notice: 'Article was successfully created.'
+    else
+      render :new
+    end
+  end
+
+  def update
+    if @article.update(article_params)
+      redirect_to @article, notice: 'Article was successfully updated.'
+    else
+      render :edit
+    end
+  end
+
+  def destroy
+    @article.destroy
+    redirect_to articles_url, notice: 'Article was successfully destroyed.'
+  end
+
+  private
+    def set_article
+      @article = Article.find(params[:id])
+    end
+
+    def article_params
+      params.require(:article).permit(:title, :slug, :short_description, :long_description, :author, :link, :link_text)
+    end
+  end
+end
