+++ app/views/layouts/_navigation.html.erb
@@ -9,6 +9,14 @@
         <li class="nav-item active">
           <%= link_to 'Articles', articles_path, class: 'nav-link' %>
         </li>
+        <li class="nav-item dropdown">
+          <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+            Admin
+          </a>
+          <div class="dropdown-menu" aria-labelledby="navbarDropdown">
+            <%= link_to 'Articles', users_articles_path, class: 'dropdown-item' %>
+          </div>
+        </li>
         <li class="nav-item">
           <a class="nav-link" href="#">About</a>
         </li>
