+++ app/views/users/articles/_sidebar.html.erb
@@ -1,0 +1,56 @@
+<!-- Sidebar Widgets Column -->
+<div class="col-md-4">
+  <!-- Search Widget -->
+  <div class="card my-4">
+    <h5 class="card-header">Search</h5>
+    <div class="card-body">
+      <div class="input-group">
+        <input type="text" class="form-control" placeholder="Search for...">
+        <span class="input-group-btn">
+          <button class="btn btn-secondary" type="button">Go!</button>
+        </span>
+      </div>
+    </div>
+  </div>
+  <!-- Categories Widget -->
+  <div class="card my-4">
+    <h5 class="card-header">Categories</h5>
+    <div class="card-body">
+      <div class="row">
+        <div class="col-lg-6">
+          <ul class="list-unstyled mb-0">
+            <li>
+              <a href="#">Web Design</a>
+            </li>
+            <li>
+              <a href="#">HTML</a>
+            </li>
+            <li>
+              <a href="#">Freebies</a>
+            </li>
+          </ul>
+        </div>
+        <div class="col-lg-6">
+          <ul class="list-unstyled mb-0">
+            <li>
+              <a href="#">JavaScript</a>
+            </li>
+            <li>
+              <a href="#">CSS</a>
+            </li>
+            <li>
+              <a href="#">Tutorials</a>
+            </li>
+          </ul>
+        </div>
+      </div>
+    </div>
+  </div>
+  <!-- Side Widget -->
+  <div class="card my-4">
+    <h5 class="card-header">About</h5>
+    <div class="card-body">
+      These are the articles that turned out to be very interesting to me
+    </div>
+  </div>
+</div>
