+++ app/views/users/articles/index.html.erb
@@ -1,0 +1,33 @@
+<!-- Page Content -->
+<div class="container articles">
+  <div class="row">
+    <div class="col-md-8">
+      <h1 class="my-4">Interesting articles</h1>
+
+      <% @articles.each do |article| %>
+        <div class="card mb-4">
+          <div class="card-body">
+            <h2 class="card-title"><%= article.title %></h2>
+            <p class="card-text"><%== article.short_description %></p>
+            <%= link_to 'Show', article, class: 'btn btn-primary' %>
+            <%= link_to 'Edit', users_article_path(article), class: 'btn btn-warning' %>
+            <%= link_to 'Delete', users_article_path(article), class: 'btn btn-danger', method: :delete, data: { confirm: 'Are you sure?' } %> 
+          </div>
+          <div class="card-footer text-muted">
+            Posted on <%= article.posted_on %> by <%= article.author %>
+          </div>
+        </div>
+      <% end %>
+      <ul class="pagination justify-content-center mb-4">
+        <li class="page-item">
+          <a class="page-link" href="#">&larr; Older</a>
+        </li>
+        <li class="page-item disabled">
+          <a class="page-link" href="#">Newer &rarr;</a>
+        </li>
+      </ul>
+    </div>
+		<%= render 'articles/sidebar' %>
+  </div>
+
+</div>
