+++ app/views/users/articles/show.html.erb
@@ -1,0 +1,22 @@
+<!-- Page Content -->
+<div class="container articles">
+	<div class="row">
+		<!-- Post Content Column -->
+		<div class="col-lg-8">
+			<!-- Title -->
+			<h1 class="mt-4"><%= @article.title %></h1>
+			<!-- Author -->
+			<p class="lead">
+				by <%= @article.author %>
+			</p>
+			<hr>
+			<p>Posted on <%= @article.posted_on %>, <%= link_to 'original article', @article.link, target: :_blank %></p>
+			<hr>
+			<!-- Post Content -->
+				<%== @article.long_description %>
+			<hr>
+		</div>
+    <%= render 'articles/sidebar' %>
+	</div>
+	<!-- /.row -->
+</div>
